+++ src/Icons/AppShellIcons.php
@@ -70,6 +70,5 @@ final class AppShellIcons
     public const DOWNLOAD = 'download';
     public const FOLDER = 'folder';
     public const FILE = 'file';
-
     public const COMMENT = 'comment';
 }